+++ android/src/main/kotlin/it/aesys/flutter_cast_video/FlutterVideoCastPlugin.kt
@@ -1,7 +1,6 @@
 package it.aesys.flutter_cast_video
 
 import androidx.annotation.NonNull;
-
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.activity.ActivityAware
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding